Crate slog_global

Global loggers for slog-rs.

Provides a set of logging macros to free user from manually passing Logger objects around.

This crate is similar to slog-scope, but is simpler and faster. Also replacing macros will be less likely to break existing code base.

Not advised to be used in libraries.

Macros

crit
Logs a critical level message using the global logger.

debug
Logs a debug level message using the global logger.

error
Logs a error level message using the global logger.

info
Logs a info level message using the global logger.

trace
Logs a trace level message using the global logger.

warn
Logs a warning level message using the global logger.

Functions

borrow_global
Temporary borrows the global Logger.

clear_global
Clears the global Logger and discard future logging.

get_global
Gets the global Logger.

redirect_std_log
Starts redirecting all logs from the log crate slog_global::borrow_global().

set_global
Sets the global Logger.